+++ model_type/__init__.py
@@ -1,2 +1,0 @@
-from .bases import *
-from .yolov8 import *