+++ model_type/yolov8/detection/ultralytics_based/__init__.py
@@ -1,2 +1,0 @@
-from .yolov8_base import *
-from .megadetectorv6 import *